requirements.txt
twitchAPI==4.5.0
Mastodon.py==2.1.4
boto3==1.40.58
hvac==2.3.0
atproto==0.0.62
doppler-sdk==1.3.0
python-dotenv==1.1.1
google-api-python-client==2.185.0
google-genai==1.46.0
# CVE-2025-4565 fix: Requires protobuf >= 4.25.8 (v4), >= 5.29.5 (v5), or >= 6.31.1 (v6)
# Using v4 for compatibility; upgrade to v5/v6 requires testing
protobuf>=4.25.8,<5.0.0
requests==2.32.5
discord.py==2.6.4
matrix-nio==0.25.2
beautifulsoup4==4.14.2